+++ com.unity.render-pipelines.high-definition/CHANGELOG.md
@@ -71,6 +71,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - Added the Tint parameter to Sky Colored Fog.
 - Added of Screen Space Reflections for Transparent materials
 - Added a fallback for ray traced area light shadows in case the material is forward or the lit mode is forward.
+- Added a new debug mode for light layers.
 
 ### Fixed
 - Update documentation of HDRISky-Backplate, precise how to have Ambient Occlusion on the Backplate